+++ src/aatoolbox/__init__.py
@@ -9,6 +9,7 @@ from aatoolbox.config.countryconfig import (
     create_custom_country_config,
 )
 from aatoolbox.datasources.codab.codab import CodAB
+from aatoolbox.datasources.fewsnet.fewsnet import FewsNet
 from aatoolbox.datasources.iri.iri_seasonal_forecast import (
     IriForecastDominant,
     IriForecastProb,
@@ -24,5 +25,6 @@ __all__ = (
     "GeoBoundingBox",
     "IriForecastProb",
     "IriForecastDominant",
+    "FewsNet",
     "__version__",
 )